feat: added cli command to run the application, added makefile to run the application
