+++ Makefile
@@ -4,6 +4,7 @@ test:
 	echo "testing application"
 run:
 	echo "running application"
+	go run cmd/crack/main.go test.csv 5
 build:
 	echo "bulding application"
 tidy:
